support .wmv, .asf formats

+++ src/observer.hpp
@@ -21,7 +21,7 @@ public:
 private:
 	void createWatcher();
 	QFileSystemWatcher* watcher;
-	std::map<QString, unsigned int> m_newVideos;
+	std::map<QString, unsigned long long int> m_newVideos;
 signals:
 	void directoryChanged(const QString&);
 	void Complete(QString);
